+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add entry banner animation
+- Add EntryBanner animation between Login and StartChallenge
 
 ## [0.14.0] - 2026-01-03